requirements.txt
kafka-python
pyspark
numpy
